refactor: replace SIGDescribe call with SIG decorator
We are currently using `SIGDescribe` funcs to generalize the decoration of
`Describe` ginkgo nodes for each SIG in order to adhere to 'dry' principle
by avoiding repetition of SIG label and decorator per each `Describe`.
This is done in all cases where there's multiple test files for a SIG,
since each of those has a `Describe` node.

Above approach breaks ginkgo outline generation, which is essential for our
CI tooling, both current and in development.
When ginkgo outline parses the AST to determine the nodes it looks up func
names matching with `Describe` exactly [1] - therefore the generated outline
is missing `SIGDescribe` nodes.
This would also happen for all other cases where funcs with names
differing from the ginkgo DSL would return a `Describe` or any other
ginkgo func.

The new approach extends the parameter list of the ginkgo node function
call with a wrapper function `SIG`, located in the `decorators` package.

This will make the `SIGDescribe` (and matching `FSIGDescribe` and
`PSIGDescribe`) obsolete, while repairing the outline tool for k/kubevirt.

Following commits will inline the `SIGDescribe` function calls, replacing
it with a direct usage of the `SIG` function. After that `SIGDescribe
will be obsolete and therefore removed. As cleanup also the `FSIGDescribe`
and `PSIGDescribe` will be removed, since it now is easier to modify the
`Describe` call itself.

[1]: https://github.com/onsi/ginkgo/blob/cbcf39a17b9a3aeea0281ce7779a42eb838334b7/ginkgo/outline/ginkgo.go#L147
