+++ tests/decorators/BUILD.bazel
@@ -2,7 +2,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = ["decorators.go"],
+    srcs = [
+        "decorators.go",
+        "sigs.go",
+    ],
     importpath = "kubevirt.io/kubevirt/tests/decorators",
     visibility = ["//visibility:public"],
     deps = ["//vendor/github.com/onsi/ginkgo/v2:go_default_library"],
