+++ tests/storage/framework.go
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *
- * Copyright 2020 Red Hat, Inc.
+ * Copyright the KubeVirt Authors.
  *
  */
 
@@ -26,13 +26,17 @@ import (
 )
 
 func SIGDescribe(text string, args ...interface{}) bool {
-	return Describe("[sig-storage] "+text, decorators.SigStorage, args)
+	return Describe(SIG(text, args))
 }
 
 func FSIGDescribe(text string, args ...interface{}) bool {
-	return FDescribe("[sig-storage] "+text, decorators.SigStorage, args)
+	return FDescribe(SIG(text, args))
 }
 
 func PSIGDescribe(text string, args ...interface{}) bool {
-	return PDescribe("[sig-storage] "+text, decorators.SigStorage, args)
+	return PDescribe(SIG(text, args))
+}
+
+func SIG(text string, args ...interface{}) (extendedText string, newArgs []interface{}) {
+	return decorators.SIG("[sig-storage]", decorators.SigStorage, text, args)
 }
